+++ lib/src/main/java/com/diffplug/spotless/rome/RomeStep.java
@@ -427,9 +427,23 @@ public class RomeStep {
 			var stdin = input.getBytes(StandardCharsets.UTF_8);
 			var args = buildBiomeCommand(file);
 			if (logger.isDebugEnabled()) {
-				logger.debug("Running Biome comand to format code: '{}'", String.join(", ", args));
+				logger.debug("Running Biome command to format code: '{}'", String.join(", ", args));
 			}
-			return runner.exec(stdin, args).assertExitZero(StandardCharsets.UTF_8);
+			var runnerResult = runner.exec(stdin, args);
+			var stdErr = runnerResult.stdErrUtf8();
+			if (stdErr != null && !stdErr.isEmpty()) {
+				logger.warn("Biome stderr ouptut for file '{}'\n{}", file, stdErr.trim());
+			}
+			var formatted = runnerResult.assertExitZero(StandardCharsets.UTF_8);
+			// When biome encounters an ignored file, it does not output any formatted code
+			// Ignored files come from (a) the biome.json configuration file and (b) from
+			// a list of hard-coded file names, such as package.json or tsconfig.json.
+			if (formatted == null || formatted.isEmpty()) {
+				return input;
+			}
+			else {
+				return formatted;
+			}
 		}
 
 		/**